requirements.txt
streamlit>=1.40.0
groq>=0.12.0
pydantic>=2.0
python-dotenv>=1.0.0
mem0ai>=0.1.0
